Update ar1.py
the function PF is now working and this produces two plots. the first is the output of the model and the second shows the approximations to the parameter. the solid blue line is the actual parameter and the dotted blue line is the least squares estimate of this. the red lines show the mean estimate from the pf and two standard deviations either way.